Rename decorators to convert images to acceptable format
